Refactor code, remove Shep.java
Some code is not of quality. Mostly, SLAP. Shep.java is also not needed
anymore after moving over to GUI.

Refactor the code to improve code quality and remove Shep.java.

Code was refactored in order to adhere to SLAP. Several classes had
multiple layers of abstraction so method extraction had to be done.