Add assertations to reinforce logic
Some methods depended on the logic of other functions. To ensure that
the logic is followed  so that it can be utilised correctly by the class
Assertations were added.

Add Assertations to methods that require other methods to follow their
logic properly. Apart fromthe Commands class assertations never replaced
if else statements.

This is because assertations shouldn't be used to control the program
flow.