Add duplicate check on Task addition
On the addition of any type of Task into a TaskList, no duplicate checks
are made, allowing the chance for a duplicate task to be added into the
test list.

This is a redundant feature and duplicate Tasks shouldn't be allowed.

In order to fix this, a checkDuplicates was added in the TaskList
class as well as an equals method in Task.

Also updated TaskListTests and CommandsTest classes to account for newer
situations.